src/uk_gov/requirements.txt
pdfplumber
pandas
python-dateutil